requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
pydantic==2.8.2
httpx==0.27.0
duckdb==1.0.0
polars==1.4.1
pandas==2.2.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy==1.26.4
python-dotenv==1.0.1 
